// packages/shared/notionable/src/types.ts
export interface BalanceSummaryResult {
  balanceSummary: BalanceSummary[];
  noteSummary: NOTESummary | null;
}

export interface AssetSummaryResult {
  assetSummary: AssetSummary[];
  tradeHistory: TradeHistory[];
}

export interface SummaryUpdateResult {
  balance: BalanceSummaryResult;
  asset: AssetSummaryResult;
}